refactor(donate): extract widget subcomponents
