+++ src/app/wallet/components/cryptoTypes.ts
@@ -1,0 +1,2 @@
+export const cryptoOptions = ['ETH', 'BTC', 'SOL', 'LTC', 'DOGE'] as const;
+export type CryptoType = typeof cryptoOptions[number];
